Lesson 5. Task 15

+++ src/main/java/students/kate_br/lesson_5/level_2/task_14/AverageArray.java
@@ -9,7 +9,7 @@ class AverageArray {
      public static void main(String[] args) {
          Random random = new Random();
 
-         double[] num = new double[3];
+         int[] num = new int[3];
 
          for (int i = 0; i < num.length; i++){
              num[i] = random.nextInt(4);
